// src/app/shared/product.interface.ts
/******************************************************************************
 * Title: product.interface.ts
 * Author: Jeff Shepherd, Clayton Stacy, Chris Bohnet
 * Modified by:
 * Date: 10/28/2020
 * Description: product interface
 *****************************************************************************/

 export interface Product {
   _id: string;
   price: number;
   text: string;
 }